Revamped code for axios. Giving referenceerror
