+++ index.html
@@ -26,6 +26,8 @@
 
     <!-- Use the api to call for a thing -->
     <nav id="mytext"></nav>
+
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
     <script src="app.js"></script>
 </body>
 </html>
