+++ package.json
@@ -2,12 +2,12 @@
     "name": "app",
     "version": "0.0.1",
     "description": "",
-    "main": "index.js",
+    "main": "app.js",
     "scripts": {},
     "author": "",
     "license": "MIT",
     "dependencies": {
-        "axiom": "^0.1.6",
+        "axiom": "^0.24.0",
         "axios": "latest",
         "body-parser": "latest",
         "ejs": "latest",
